fix: Change Energy to Big Int
* Integer is not large enough to hold enough digits.
* Increase worker count to 2
* Remove duplicated column on Model

# Conflicts:
#	backend/lcfs/db/models/compliance/FuelSupply.py

# Conflicts:
#	backend/lcfs/db/models/compliance/FuelSupply.py

+++ backend/lcfs/db/migrations/versions/2024-10-18-19-46_a731a32947dc.py
@@ -135,7 +135,7 @@ def upgrade() -> None:
         "fuel_supply",
         "energy",
         existing_type=sa.DOUBLE_PRECISION(precision=53),
-        type_=sa.Integer(),
+        type_=sa.BigInteger(),
         existing_comment="Energy content of the fuel supplied",
         existing_nullable=True,
     )
